Refactoring of the AccountInfo, Role and JWT/Login logic

+++ spring-boot-backend/src/main/java/com/eafc/springbootbackend/controllers/administrator/AdminCategoryController.java
@@ -17,7 +17,7 @@ public class AdminCategoryController {
     }
 
     @GetMapping("getCategoryById")
-    public Category getCategoryById(@RequestParam("catId") Integer categoryId) {
+    public Category getCategoryById(@RequestParam("catId") Long categoryId) {
         return categoryService.findCategoryById(categoryId).get();
     }
