+++ spring-boot-backend/src/main/java/com/eafc/springbootbackend/entities/product/Stock.java
@@ -15,7 +15,7 @@ public class Stock {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int stockId;
+    private Long stockId;
 
     @NotNull
     @NotBlank
