+++ spring-boot-backend/src/main/java/com/eafc/springbootbackend/entities/shopping/CartItem.java
@@ -12,7 +12,7 @@ public class CartItem {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int cartItemId;
+    private Long cartItemId;
 
     @ManyToOne
     @NotNull
