+++ spring-boot-backend/src/main/java/com/eafc/springbootbackend/services/product/CategoryServiceImpl.java
@@ -18,7 +18,7 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public Optional<Category> findCategoryById(int categoryId) {
+    public Optional<Category> findCategoryById(Long categoryId) {
         return categoryRepository.findById(categoryId);
     }
 
@@ -33,7 +33,7 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public void deleteCategory(int categoryId) {
+    public void deleteCategory(Long categoryId) {
         categoryRepository.deleteById(categoryId);
     }
 }
